requirements.txt
# Core dependencies
mcp>=1.0.0
pydantic>=2.0.0
asyncio-mqtt>=0.16.0
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.0.0
typing-extensions>=4.8.0

# System monitoring and control
psutil>=5.9.8
pyautogui>=0.9.54
pillow>=10.0.0

# Windows-specific dependencies
pywin32>=306; sys_platform == 'win32'
wmi>=1.5.1; sys_platform == 'win32'

# Security
cryptography>=42.0.0
bcrypt>=4.0.0

# Monitoring and metrics
prometheus-client>=0.17.0
watchdog>=3.0.0

# Configuration
pyyaml>=6.0.0
jsonschema>=4.19.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0

# Code quality
flake8>=6.0.0
black>=23.0.0
mypy>=1.5.0
isort>=5.12.0

# Additional utilities
aiohttp>=3.8.0
click>=8.1.0
